+++ lib/karafka/messages.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module Karafka
+  # Messages namespace encapsulating all the logic that is directly related to messages handling
+  module Messages
+  end
+end