+++ lib/karafka/pro/active_job.rb
@@ -13,14 +13,8 @@
 
 module Karafka
   module Pro
-    module Routing
-      module Features
-        # Long-Running Jobs feature config and DSL namespace.
-        #
-        # Long-Running Jobs allow you to run Karafka jobs beyond `max.poll.interval.ms`
-        class LongRunningJob < Base
-        end
-      end
+    # Karafka Pro ActiveJob components
+    module ActiveJob
     end
   end
 end