+++ lib/karafka/pro/encryption/contracts/config.rb
@@ -14,7 +14,6 @@
 module Karafka
   module Pro
     module Encryption
-      # Encryption related contracts
       module Contracts
         # Makes sure, all the expected config is defined as it should be
         class Config < ::Karafka::Contracts::Base