+++ lib/karafka/pro/encryption/setup.rb
@@ -13,13 +13,9 @@
 
 module Karafka
   module Pro
-    module Routing
-      module Features
-        # Long-Running Jobs feature config and DSL namespace.
-        #
-        # Long-Running Jobs allow you to run Karafka jobs beyond `max.poll.interval.ms`
-        class LongRunningJob < Base
-        end
+    module Encryption
+      # Setup and config related encryption components
+      module Setup
       end
     end
   end