+++ lib/karafka/pro/processing/jobs.rb
@@ -13,13 +13,10 @@
 
 module Karafka
   module Pro
-    module Routing
-      module Features
-        # Long-Running Jobs feature config and DSL namespace.
-        #
-        # Long-Running Jobs allow you to run Karafka jobs beyond `max.poll.interval.ms`
-        class LongRunningJob < Base
-        end
+    # Pro components related to processing part of Karafka
+    module Processing
+      # Pro jobs
+      module Jobs
       end
     end
   end