+++ lib/karafka/pro/routing.rb
@@ -13,14 +13,8 @@
 
 module Karafka
   module Pro
+    # Namespace for Pro routing enhancements
     module Routing
-      module Features
-        # Long-Running Jobs feature config and DSL namespace.
-        #
-        # Long-Running Jobs allow you to run Karafka jobs beyond `max.poll.interval.ms`
-        class LongRunningJob < Base
-        end
-      end
     end
   end
 end