+++ lib/karafka/pro/routing/features.rb
@@ -14,12 +14,8 @@
 module Karafka
   module Pro
     module Routing
+      # Namespace for additional Pro features
       module Features
-        # Long-Running Jobs feature config and DSL namespace.
-        #
-        # Long-Running Jobs allow you to run Karafka jobs beyond `max.poll.interval.ms`
-        class LongRunningJob < Base
-        end
       end
     end
   end