+++ lib/karafka/time_trackers.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+module Karafka
+  # Time trackers module.
+  #
+  # Time trackers are used to track time in context of having a time poll (amount of time
+  # available for processing) or a pausing engine (pause for a time period).
+  module TimeTrackers
+  end
+end